Fixes for WP display and centralization of notification code
Fixes for WP display:
1) Fixed the bearing value display
2) Made display of distance more accurate

A new class, CommonEntity, is introduced. It centralizes code that was previously duplicated in DefaultFolder, DefaultEncoder and DefaultInput:
1) Code implementing the INotifier interface.
2) Registering a binding with a certain key and initial value and adding it to the collection used for notification.
3) The collection of bindings in a folder, encoder or input.

Apart from that a lot of refactoring in folders: They now all use meaningful names of bindings rather than indexing in a list. The array still exists but is now hidden in CommonEntity. Similar refactorization of encoders and inputs will be made soon and then the binding collection in CommonEntity will be made private.